[CI] Enable docker build on native riscv64 runner

+++ .ci/docker/common/install_base.sh
@@ -61,6 +61,7 @@ install_ubuntu() {
     libasound2-dev \
     libsndfile-dev \
     ${maybe_libomp_dev} \
+    python-is-python3 \
     software-properties-common \
     wget \
     sudo \
@@ -137,15 +138,20 @@ case "$ID" in
     ;;
 esac
 
-# Install Valgrind separately since the apt-get version is too old.
-mkdir valgrind_build && cd valgrind_build
-VALGRIND_VERSION=3.20.0
-wget https://ossci-linux.s3.amazonaws.com/valgrind-${VALGRIND_VERSION}.tar.bz2
-tar -xjf valgrind-${VALGRIND_VERSION}.tar.bz2
-cd valgrind-${VALGRIND_VERSION}
-./configure --prefix=/usr/local
-make -j$[$(nproc) - 2]
-sudo make install
-cd ../../
-rm -rf valgrind_build
-alias valgrind="/usr/local/bin/valgrind"
+if [[ $(uname -m) == "riscv64" ]]; then
+  # Support was added in valgrind-3.26; Skip it
+  true
+else
+  # Install Valgrind separately since the apt-get version is too old.
+  mkdir valgrind_build && cd valgrind_build
+  VALGRIND_VERSION=3.20.0
+  wget https://ossci-linux.s3.amazonaws.com/valgrind-${VALGRIND_VERSION}.tar.bz2
+  tar -xjf valgrind-${VALGRIND_VERSION}.tar.bz2
+  cd valgrind-${VALGRIND_VERSION}
+  ./configure --prefix=/usr/local
+  make -j$[$(nproc) - 2]
+  sudo make install
+  cd ../../
+  rm -rf valgrind_build
+  alias valgrind="/usr/local/bin/valgrind"
+fi
